pip_library: test for presence of --system option before invoking pip
The version of pip provided by `python3-pip` in Debian 12 will drop the
custom `--system` option - passing it results in pip exiting with an
error. Rather than always passing `--system` if the host OS is
Debian-like, test whether the configured pip tool implements a
`--system` option, and if so, pass it to the real pip invocation.

Fixes #81.