Upgrade importlib-metadata to v6.8.0
The version of importlib-metadata in the bootstrap pex masks all other
versions, and v1.5.0 is pretty old now - it's missing functionality
required by other modules on PyPi, such as `EntryPoints`. Bump it to the
latest version, v6.8.0.

The `__version__` patch is no longer needed because importlib-metadata
stopped presenting a `__version__` attribute in v2.0.0.