Provide raw wheel file when `py_whl` is required (#134)
It's often useful to be able to get the untouched wheel downloaded by
`python_wheel` rather than a Please-ified pex file containing the
package's code, e.g. when the `python_wheel` target is a dependency of a
package that is being built using a non-Please build system like pip.
Provide the wheel when `py_whl` is required by a downstream target.