Make in-repo interpreter available to test at runtime (#141)
Currently if you pass an `interpreter` to `python_test` which is the output of a build target, then when run remotely (so that the shebang of the output pex references `/tmp/plz_sandbox` instead of a directory in the local `plz-out`), the test errors because it can't find the interpreter. For example:

```python
python_test(
    name = "foo_test",
    srcs = ["foo_test.py"],
    interpreter = "//third_party/python3/cpython:cpython_3.10|python",
)
```

results in 

```
bash: /tmp/plz_sandbox/foo_test.pex: /tmp/plz_sandbox/third_party/python3/cpython/usr/python3.10/bin/python3: bad interpreter: No such file or directory
```

This change modifies `python_test` to check whether `interpreter` looks like a build label, and if so pass it to the final pex rule as `data`.