Add more defaults to interpreter to handle generic wheels.
The interpreter being locked to the current interpreter in the passed
arguments was too restrictive. We needed to account for two things:

1. Intepreters should be a set including generic ones (such as py3)
2. The platform should always match the short phrase any

With these two changes made we now verify correctly that the following
command works:

```
plz run //tools/wheel_resolver -- --package pytest_unordered --version 0.6.0
```

Includes a test case of the above to ensure no regressions.