Remove dependency on importlib-metadata in meta path finder (#176)
importlib-metadata is a third-party module that isn't included in the
bootstrap pex, preventing the `ModuleDirImport` class from acting as a
meta path finder unless it happens to be included as a dependency of
downstream code. Depend on `importlib.metadata` instead, which has been
part of the standard library since 3.8.

Fixes #175.