Load distribution metadata for unimported modules in meta path finder (#178)
`PexDistribution` is only used to locate distribution metadata for
modules that already exist in `sys.modules`, requiring a module to be
imported by the interpreter before its distribution metadata can be
queried. Remove this restriction, allowing distribution metadata to be
queried at any time provided the module exists in the pex file.

Fixes #177.