Handle empty contexts in meta path finder (#184)
Implementations of `importlib.abc.MetaPathFinder` are required to return
an iterable of all `Distribution`s when an empty context is given, but
`ModuleDirImport` is currently only capable of returning a single
`PexDistribution` when a name is given in the context (and raises an
exception otherwise). Pre-compute a mapping of distribution names to
lists of distribution packages with those names that exist in the pex
file at initialisation time, and return all or a subset of them when
`find_distributions` is called.