Run CI tests against in-repo pex tool (#185)
The CI tests currently only run against the released version of
please_pex defined in `TOOLS_VERSION` in `tools/BUILD`, which means that
changes made in please_pex PRs aren't actually tested in CI until a new
tools release has already been cut and the version number bumped in
`TOOLS_VERSION`. Add a new dimension to the GHA matrix that builds the
in-repo version of please_pex and overrides the value of `PexTool` in
the plugin config so that tests are run against it too.