CI: use interpreter provided by `actions/setup-python`
`actions/setup-python` "installs" a non-standard Python interpreter in
the CI environment by adding a directory containing it to `PATH`, but
the updated `PATH` isn't inherited by the Please build sandbox, so the
system-wide Python interpreter is always used. Pass through the value of
`PATH` to the sandbox so that the correct `actions/setup-python`
interpreter is used.

Fixes #127.