CI: use interpreter provided by `actions/setup-python` (#192)
`actions/setup-python` "installs" a non-standard Python interpreter in
the CI environment by adding a directory containing it to `PATH`, but
the updated `PATH` isn't inherited by the Please build sandbox, so the
system-wide Python interpreter is always used. Set the path to `python3`
provided by `actions/setup-python` as the default interpreter used by
the plugin in CI environments.

Fixes #127.