Remove NumPy from test suite
NumPy's Python version support policy has become increasingly narrower,
to the extent that it no longer supports all supported versions of
Python (which python-rules still does). It's become untenable to use it
for testing `pip_library`. Remove it, and all tests and third-party
packages that depend on it. For the purposes of the zip-unsafe labels
test, replace it with PyYAML, which still has broad compatibility across
Python versions.