Remove site-wide dependency on setuptools in test suite
One set of tests, in `//test/local`, ensures that setuptools can be
imported when Python is not invoked with the `-S` option and that it
can't be imported when Python is invoked with the `-S` option. This
assumes that setuptools is installed site-wide (i.e., in a directory
added to `sys.path` by the site module), which isn't necessarily the
case.

Refactor these tests to check the value of `sys.flags.no_site`, which is
an explicit way to identify whether the `-S` option was passed to
Python.