Implement missing `locate_file` method in pex meta path finder
This is an abstract method of `importlib.metadata.Distribution` that
must be implemented in Python >= 3.12 to avoid `DeprecationWarning`s. It
doesn't actually make sense for `locate_file` to return anything in a
`PexDistribution`, given that the files it finds exist in a zip file
rather than on a real file system.