Really implement `locate_file` method in pex meta path finder (#208)
The importlib.metadata documentation in Python 3.13 implies that it's
fine for a `Distribution`'s implementation of the `locate_file` abstract
method to raise an error if the distribution doesn't exist as real files
on a real file system, but this error will bubble up when the `files`
method is called (because `files` gets a `pathlib.Path`-like object from
`locate_file` for each file in `RECORD` and then checks that the path
exists by calling `exists()` on each return value).

In `locate_file`, construct and return a `zipfile.Path` object pointing
to the given file in the distribution in the pex file. `zipfile.Path`
takes care of the path existence check itself, so `files` works as
expected when running under Python >= 3.12.