Formalise Python version support policy
In `README.md`, state that we aim to support every Python branch still
supported by the PSF, and state the plugin versions that added or
removed support for specific branches. The plugin versions listed here
are based on when each Python version gained native-speed coverage
tracing in please_pex, which is a reasonable approximation of when that
version was supported by the plugin.