Bump grpcio to 1.75.1
grpcio is only used as part of a `zip-unsafe` label e2e test, but
there's no pre-built Python 3.14-compatible wheel for 1.67.1 on PyPI so
`pip_library` is forced to build the module from source as a dependency
of that test. Bump the version to 1.75.1, which has pre-built binary
wheels for every version of Python this plugin currently supports.