Make `pip_library` zip safety end-to-end tests self-contained
The end-to-end tests that check whether `pip_library` correctly applies
the `py:zip-unsafe` label to a zip-unsafe target depend on two
third-party Python modules, PyYAML and grpcio. The pre-built wheels on
PyPI are architecture-specific, so their versions have to be kept in
line with ones for which pre-built wheels are available on PyPI
(otherwise pip will take > 10 minutes and a lot of resources to build
them from source); even when a pre-built wheel is available for the
SOABI in question, the download is disproportionately large for such a
simple test (7-12MB for grpcio, depending on the SOABI).

Replace the `pip_library` zip safety tests with a single, fully
self-contained e2e test that occurs within an e2e-specific Please repo,
requiring no downloads from PyPI.