CI: simplify steps for using in-repo please_pex tool
The circular dependency between please_pex and its bootstrap archive
was broken in #266, allowing for the plugin to be configured to use
`//tools_please_pex` as its own pex tool. This is simpler than building
the please_pex binary in one step, copying it elsewhere in the file
system for persistence, then configuring future steps to use that binary
as the pex tool.