Remove cx-Oracle import test (#271)
`//test:cx_oracle_darwin_build_test` just ensures that cx-Oracle is
importable and defines a `__version__` attribute. The test name implies
that the build process for cx-Oracle itself - specifically on Darwin - is
what matters in this test, although it's not clear why: the test
predates Please's plugin architecture, and was inherited from the main
Please repo.

cx-Oracle has been abandoned since 2022 [1] and hasn't had a new version
published on PyPI since 2021 [2]. The final version only has pre-built
wheels available for Python <= 3.10 manylinux SOABIs, so pip has to
build it from source for the vast majority of SOABIs. It's not clear why
this is a price worth paying any more, so remove it (and the test that
depends on it).

[1] https://oracle.github.io/python-cx_Oracle/
[2] https://pypi.org/project/cx-Oracle/8.3.0/