Define in-repo Python interpreters for all supported platforms
Including, for the first time, FreeBSD. This involves a bit of a kludge:
python-build-standalone doesn't produce FreeBSD-compatible releases, so
we have to manually download and unpack the Ports packages and patch the
binaries so they behave like standalone interpreters.

This will allow us to run the tests on all platforms as part of CI
testing and produce FreeBSD-compatible versions of please_pex when the
time comes.