Allow flag to appear anywhere in arguments list in `flags` test
The flag passed in the `flags` parameter of `//test:flags_test` isn't
always guaranteed to be the first argument in the interpreter's
arguments list - other factors, e.g. overriding the value of the
plugin's `InterpreterOptions` configuration option when running the
test, can affect the argument order. Relax the assertion in the test so
it passes if the flag appears anywhere in the arguments list.