Upgrade pytest to 9.0.1
Lots of interesting new features and bug fixes have been added to pytest
since v7, including support for Python 3.13 and 3.14 - see the
[changelog](https://docs.pytest.org/en/stable/changelog.html). There are
many breaking changes, though, so it'd be best to align this upgrade
with a future major version of the plugin - luckily, we're about to
release one.

This version drops one dependency (colorama) and introduces two new ones
(pygments, typing-extensions).