Remove `InterpreterOptions` plugin configuration option
The documentation for the `InterpreterOptions` plugin configuration
option is ambiguous - it implies the options within are used by Python
while the .pex is being built (e.g. while compiling bytecode), but
actually they're embedded in the .pex and used at run time. Now they
aren't used at all: this functionality is instead provided by
`RuntimeInterpreterArgs`.

Remove `InterpreterOptions`, since it is obsolete. Also tidy up some of
the variable names in the Python build defs to make it explicit that the
arguments in `RuntimeInterpreterArgs` are in fact passed to Python at
run time.