Build please_pex releases in native environments (#287)
Now that please_pex writes a preamble C binary that is native to each
platform supported by Please, please_pex is no longer just a Go binary
that can be cross-compiled for every platform on a Linux amd64 runner:
the preamble needs to be built natively for each platform and embedded
in the please_pex binary for that platform.

Reorganise the please_pex test, build and release workflow so that the
release binaries are built natively on their respective platforms:
[alpine-builder](https://github.com/please-build/alpine-builder) for
Linux (linking statically to musl results in a nice ~94KB binary),
[freebsd-builder](https://github.com/please-build/freebsd-builder) for
FreeBSD, and GitHub Actions' macOS 15 runners for Darwin.

Although Ubuntu arm64 GitHub runners are available, the linux_arm64
release has to be built on an Ubuntu amd64 runner because
JavaScript-based workflow actions currently can't be executed on an
Alpine arm64 container running on an arm64 host - see:

- https://github.com/actions/upload-artifact/issues/739
- https://github.com/actions/runner/issues/801