Preamble: open .pex file from its canonical path (#292)
`argv[0]` might look like a relative path without actually existing
relative to the current working directory - this is the case if (e.g.)
the .pex file has been copied into a directory in `PATH` and the shell's
`PATH` search behaviour is being relied upon to locate the .pex file.
Get the canonical path to `argv[0]` before attempting to open it and
read the preamble configuration within it.